+++ src/sphere.cpp
@@ -3,7 +3,7 @@
 namespace PathTrace
 {
 
-Sphere::Sphere(Vector3D center, double r, const Material * material)
+Sphere::Sphere(Vector3D center, float r, const Material * material)
 {
     this->center = center;
     this->r = r;
@@ -21,7 +21,7 @@ namespace
 class SphereSpanIterator : public SpanIterator
 {
 public:
-    SphereSpanIterator(const Vector3D & center, double r_squared, const Material * material)
+    SphereSpanIterator(const Vector3D & center, float r_squared, const Material * material)
         : center(center), r_squared(r_squared)
     {
         theSpan.startMaterial = material;
@@ -32,15 +32,15 @@ public:
     {
         ended = false;
         Vector3D origin_minus_center = ray.origin - center;
-        double b = dot(origin_minus_center, ray.dir);
-        double c = dot(origin_minus_center, origin_minus_center) - r_squared;
-        double sqrt_arg = b * b - c;
+        float b = dot(origin_minus_center, ray.dir);
+        float c = dot(origin_minus_center, origin_minus_center) - r_squared;
+        float sqrt_arg = b * b - c;
         if(sqrt_arg <= eps)
         {
             ended = true;
             return;
         }
-        double sqrt_v = std::sqrt(sqrt_arg);
+        float sqrt_v = std::sqrt(sqrt_arg);
         theSpan.start = -b - sqrt_v;
         theSpan.end = -b + sqrt_v;
         theSpan.startNormal = normalize(ray.getPoint(theSpan.start) - center);
@@ -70,7 +70,7 @@ private:
     Span theSpan;
     bool ended;
     const Vector3D center;
-    const double r_squared;
+    const float r_squared;
 };
 }
